docs: record typed accessor refactor
